fix for inserting a new note with a check-in and display the list of notes

+++ app/src/main/java/it/antedesk/mytrips/database/dao/ActivityDao.java
@@ -12,25 +12,25 @@ import it.antedesk.mytrips.model.CheckIn;
 public interface ActivityDao extends BaseDao<Activity> {
 
     @Query("SELECT * FROM activities WHERE id=:id")
-    Activity retrieveActivityById(int id);
+    Activity retrieveActivityById(long id);
 
     @Query("SELECT * FROM activities WHERE diary_id=:diaryId")
-    List<Activity> retrieveActivitiesByDiaryId(final int diaryId);
+    List<Activity> retrieveActivitiesByDiaryId(final long diaryId);
 
     @Query("SELECT * FROM check_ins WHERE id=:checkInId")
-    CheckIn retrieveCheckInById(final int checkInId);
+    CheckIn retrieveCheckInById(final long checkInId);
 
     @Query("SELECT * FROM activities WHERE diary_id=:id ORDER BY date_time DESC")
-    List<Activity> loadActivitiesByDiaryId(int id);
+    List<Activity> loadActivitiesByDiaryId(long id);
 
     @Query("SELECT COUNT(DISTINCT address) " +
             "FROM activities LEFT JOIN check_ins on activities.check_in_id = check_ins.id " +
             "WHERE diary_id=:diaryId " +
             "ORDER BY date_time DESC")
-    int getTotalCheckinsByDiaryId(int diaryId);
+    long getTotalCheckinsByDiaryId(long diaryId);
 
     @Query("SELECT SUM(budget) FROM activities WHERE diary_id=:diaryId")
-    double getTotalBudgetByDiaryId(int diaryId);
+    double getTotalBudgetByDiaryId(long diaryId);
 
     @Query("SELECT SUM(budget) FROM activities")
     double getTotalBudget4AllActivities();
@@ -39,5 +39,5 @@ public interface ActivityDao extends BaseDao<Activity> {
     double getTotalBudgetByCategories();
 
     @Query("SELECT category, SUM(budget) FROM activities WHERE diary_id=:diaryId GROUP BY category")
-    double getTotalBudgetByCategoriesAndDiaryId(int diaryId);
+    double getTotalBudgetByCategoriesAndDiaryId(long diaryId);
 }
