+++ app/src/main/java/it/antedesk/mytrips/database/dao/UserDao.java
@@ -9,6 +9,6 @@ import it.antedesk.mytrips.model.User;
 public interface UserDao extends BaseDao<User>{
 
     @Query("SELECT * FROM users WHERE id=:id")
-    User retrieveUserById(int id);
+    User retrieveUserById(long id);
 
 }
